third_party/WebKit/LayoutTests/http/tests/loading/sxg/sxg-location-fragment.html

<!DOCTYPE html>
<title>Navigating to a Signed Exchange should preserve URL fragment</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  await waitUntilDidFinishLoadForFrame;

  const message = await openSXGInIframeAndWaitForMessage(t, 'resources/sxg-location.sxg#fragment');
  assert_equals(message.location, 'https://127.0.0.1:8443/loading/sxg/resources/inner-url.html#fragment');
  assert_false(message.is_fallback);
}, 'Navigating to a Signed Exchange should preserve URL fragment');

</script>
</body>
